requirements.txt
boto3==1.36.16
python-dotenv==1.0.1